% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect a lazy Mongo query}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{A \code{tbl_mongo} object.}

\item{...}{Additional arguments forwarded to the executor.}
}
\value{
A tibble.
}
\description{
Collect a lazy Mongo query
}
