% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-mutate.R
\name{mutate.tbl_mongo}
\alias{mutate.tbl_mongo}
\title{Add computed fields to a lazy Mongo query}
\usage{
\method{mutate}{tbl_mongo}(.data, ...)
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Named scalar expressions.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Add computed fields to a lazy Mongo query
}
