% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs-arrange.R
\name{slice_head.tbl_mongo}
\alias{slice_head.tbl_mongo}
\title{Limit a lazy Mongo query}
\usage{
\method{slice_head}{tbl_mongo}(.data, ..., n = NULL, prop = NULL, by = NULL)
}
\arguments{
\item{.data}{A \code{tbl_mongo} object.}

\item{...}{Must be empty.}

\item{n}{Number of rows to keep.}

\item{prop}{Unsupported.}

\item{by}{Unsupported.}
}
\value{
A modified \code{tbl_mongo} object.
}
\description{
Limit a lazy Mongo query
}
